I really need to stop uploading via the website
It never ends well and then I have to use git anyway. Le sigh...
